timetable_solver/requirements.txt
ortools>=9.0
pandas>=1.0
matplotlib>=3.0
streamlit>=1.0
python-dotenv>=1.0
PyYAML>=6.0
